Remove -AsString from call to Find-DHCPServerv4LeaseIPAddres
Find-DHCPServerv4LeaseIPAddres is having inconsistent results, some
times the IPAddress propperty of the object returned from
Get-DhcpServerv4Lease is a string and sometimes it is an IPAddress
object that has an IPAddressAsString property. This will likely not be
the last time this needs to be changed as something underlying this is
not working consitently right.